// src/swagger/model/addBatteryDto.ts
/**
 * fms
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface AddBatteryDto { 
    batterySerialNumber?: string;
    batteryMake?: string;
    batteryModel?: string;
    batteryType?: string;
    batteryVoltage?: number;
    batteryAmp?: number;
    warrantyMonths?: number;
    cost?: number;
    invoiceDate?: Date;
    vendorId?: string;
    invoiceNumber?: string;
}